Move mirror logic to main ts file
